Reorganizes schema files
