+++ validation/configuration/waterSupplySchema.ts
@@ -29,26 +29,24 @@ export const waterTypes2: SelectOption[] = generateSelectOptionsFromZodEnum(
   ["No selezione", ...commonWaterTypeLabels]
 );
 
-export const boosterPumpEnum = z.enum(["NO_SELECTION", "1.5KW", "2.2KW"]);
+export const BoosterPumpEnum = z.enum(["NO_SELECTION", "1.5KW", "2.2KW"]);
 export const boosterPumps: SelectOption[] = generateSelectOptionsFromZodEnum(
-  boosterPumpEnum,
+  BoosterPumpEnum,
   ["No selezione", "1.5KW", "2.2KW"]
 );
 
 export const waterSupplySchema = z.object({
   water_type_1: WaterType1Enum,
-  booster_pump_1: boosterPumpEnum
-    .or(emptyStringOrUndefined())
-    .transform((val) =>
-      val === boosterPumpEnum.enum.NO_SELECTION || !val ? undefined : val
-    ),
+  booster_pump_1: BoosterPumpEnum.or(emptyStringOrUndefined()).transform(
+    (val) =>
+      val === BoosterPumpEnum.enum.NO_SELECTION || !val ? undefined : val
+  ),
   water_type_2: WaterType2Enum.optional().transform((val) =>
     val === WaterType2Enum.enum.NO_SELECTION ? undefined : val
   ),
-  booster_pump_2: boosterPumpEnum
-    .or(emptyStringOrUndefined())
-    .transform((val) =>
-      val === boosterPumpEnum.enum.NO_SELECTION || !val ? undefined : val
-    ),
+  booster_pump_2: BoosterPumpEnum.or(emptyStringOrUndefined()).transform(
+    (val) =>
+      val === BoosterPumpEnum.enum.NO_SELECTION || !val ? undefined : val
+  ),
   has_antifreeze: z.boolean().default(false),
 });
